Refatorando codigo.
